ci: use pyenv virtualenv-delete instead of pyenv uninstall
Updating Makefile to use pyenv virtualenv-delete to delete the virtual
environment instead of pyenv uninstall. Also, I think the hooks are
properly being created in pyenv now, so we don't need to install black,
pytest, and pdbpp outside of poetry. We can run poetry install to get
all of the pieces installed.