feat: support sending additional_flags and command to endpoints
When using the /client/proxy endpoint, users can send additional command
line options to the proxy command. In the call to `...client()`, set the
`additional_flags` parameter to a list of flags to be appended to the
default command.

Example:
```
client.client(
    additional_flags=["--opt1", "value1", "--opt2", "value2"]
)
```

When using the /client/command endpoint, users can set the command to
run. In the call to `...client()`, set the `command` parameter to a list
of version of the command to run.

Example:
```
client.client(command=["sleep", "200000000000"])
```

Added test cases for the new features.